% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mcmc.R
\name{assign_cluster}
\alias{assign_cluster}
\title{Assign clusters according to membership probabilities, for plot_D_heatmap()}
\usage{
assign_cluster(df, misc)
}
\arguments{
\item{df}{data frame with 3 variables: group, id_nogroup, mean}

\item{misc}{character strings representing the miscellaneous groups}
}
\description{
Assign clusters according to membership probabilities, for plot_D_heatmap()
}
