% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{detect_arrange}
\alias{detect_arrange}
\title{Detect community for a graph}
\usage{
detect_arrange(g, f)
}
\arguments{
\item{g}{an igraph object}

\item{f}{a community detection function}
}
\value{
a data fram with group memberships and associated numerical id
}
\description{
Detect community for a graph
}
