% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggs_list}
\alias{ggs_list}
\title{From ggs to a list of gg objects}
\usage{
ggs_list(df, density = TRUE)
}
\arguments{
\item{df}{a data frame & ggs object}

\item{density}{boolean (default to TRUE); should the density be plotted instead of the histogram}
}
\description{
From ggs to a list of gg objects
}
