% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mcmc.R
\name{graph_colour_size}
\alias{graph_colour_size}
\title{Attach colours and sizes to vertices}
\usage{
graph_colour_size(df, g)
}
\arguments{
\item{df}{data frame with memberships, id_nogroup as id, & entropy}

\item{g}{igraph object}
}
\description{
Attach colours and sizes to vertices
}
