% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{graph_to_edgelist}
\alias{graph_to_edgelist}
\title{Extract edge list of a graph object.}
\usage{
graph_to_edgelist(g)
}
\arguments{
\item{g}{an igraph graph object.}
}
\value{
An |E| by 2 tibble, where |E| is the number of edges in g.
}
\description{
Extract edge list of a graph object.
}
\examples{
g0 <- igraph::graph_from_data_frame(data.frame(a = c("1", "2"), b = c("2", "3")))
graph_to_edgelist(g0)
}
