% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{graph_to_nodelist}
\alias{graph_to_nodelist}
\title{Extract node list of a graph object.}
\usage{
graph_to_nodelist(g)
}
\arguments{
\item{g}{An igraph graph object.}
}
\value{
An |N| by 1 tibble, where |N| is the number of nodes in g.
}
\description{
Extract node list of a graph object.
}
\examples{
g0 <- igraph::graph_from_data_frame(data.frame(a = c("1", "2"), b = c("2", "3")))
graph_to_nodelist(g0)
}
