% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{list_descendants}
\alias{list_descendants}
\title{Compute list of descendants of id & loop indeices, if any}
\usage{
list_descendants(id, edgelist, max.level = 50L)
}
\arguments{
\item{id}{a character string which is a vertex in the graph}

\item{edgelist}{a data frame representing the edgelist of the graph}

\item{max.level}{maximum level of descendants to compute; default to 50}
}
\value{
a list of descendants and loop indices
}
\description{
Compute list of descendants of id & loop indeices, if any
}
