% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{matrix_to_ggs_C}
\alias{matrix_to_ggs_C}
\title{From group-to-group probabilities to ggs object}
\usage{
matrix_to_ggs_C(m_C, col = 1L)
}
\arguments{
\item{m_C}{matrix of group-to-group probabilities}

\item{col}{column to select}
}
\description{
From group-to-group probabilities to ggs object
}
