% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{matrix_to_ggs_D}
\alias{matrix_to_ggs_D}
\title{From memberships to ggs object}
\usage{
matrix_to_ggs_D(m_D, prefix = "abfx08")
}
\arguments{
\item{m_D}{matrix of membership probabilities}

\item{prefix}{prefix or id to select from}
}
\description{
From memberships to ggs object
}
