% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mcmc.R
\name{mcmc_summary_all}
\alias{mcmc_summary_all}
\title{Summary of a list of MCMC output}
\usage{
mcmc_summary_all(l)
}
\arguments{
\item{l}{list; essentially an MCMC output}
}
\description{
Summary of a list of MCMC output
}
