% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{names_in_group}
\alias{names_in_group}
\title{Extract id of a group}
\usage{
names_in_group(i, df.com)
}
\arguments{
\item{i}{integer index representing the group}

\item{df.com}{output of detect_arrange}
}
\description{
Extract id of a group
}
