% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{null_to_character_na}
\alias{null_to_character_na}
\title{From NULL to character NA}
\usage{
null_to_character_na(x)
}
\arguments{
\item{x}{vector of characters}
}
\description{
From NULL to character NA
}
