% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_C_raster}
\alias{plot_C_raster}
\title{Raster plot of group-to-group probabilities}
\usage{
plot_C_raster(df, size = 10)
}
\arguments{
\item{df}{a data frame with variables mean, row, and col}

\item{size}{size of text in ggplot2::geom_text(); default to 10}
}
\description{
Raster plot of group-to-group probabilities
}
