% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_D_heatmap}
\alias{plot_D_heatmap}
\title{Plot the membership heatmaps of all nodes}
\usage{
plot_D_heatmap(df0, df1)
}
\arguments{
\item{df0}{an output of mcmc_summary_all()}

\item{df1}{an output of assign_cluster()}
}
\description{
Plot the membership heatmaps of all nodes
}
