% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_i_heatmap}
\alias{plot_i_heatmap}
\title{Plot the positions of all nodes in topo. order}
\usage{
plot_i_heatmap(df)
}
\arguments{
\item{df}{an output of mcmc_summary_all()}
}
\description{
Plot the positions of all nodes in topo. order
}
