% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_i_vs_year}
\alias{plot_i_vs_year}
\title{Plot the (mean) positions of nodes against year}
\usage{
plot_i_vs_year(df0, df1)
}
\arguments{
\item{df0}{an output of mcmc_summary_all()}

\item{df1}{df0.link in draft.Rnw}
}
\description{
Plot the (mean) positions of nodes against year
}
