% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_network}
\alias{plot_network}
\title{Plot an igraph object with specific arguments}
\usage{
plot_network(g, seed = 110L, ...)
}
\arguments{
\item{g}{an igraph object}

\item{seed}{a number for initial seed, interpreted as a positive integer}

\item{...}{any other arguments passed to graphics::plot()}
}
\description{
Plot an igraph object with specific arguments
}
