% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorder_once}
\alias{reorder_once}
\title{Reorder rows and columns of a square matrix}
\usage{
reorder_once(Y, o)
}
\arguments{
\item{Y}{a square matrix}

\item{o}{a permutation vector of 0,1,...,(n-1), where n is the number of rows/columns of Y}
}
\value{
the reordered square matrix of same size as Y
}
\description{
Reorder rows and columns of a square matrix
}
