% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgs_mmsbm}
\alias{rgs_mmsbm}
\title{Run the regular Gibbs sampler of MMSBM}
\usage{
rgs_mmsbm(Y, o_init, alpha_init, s_alpha_init, a, b, A, B, n_swap = 1L,
  N = 1000L, thin = 1L, burnin = 100L, print_freq = 100L)
}
\arguments{
\item{Y}{the adjacency matrix representing the graph}

\item{o_init}{initial value of the ordering vector}

\item{alpha_init}{initial value of alpha}

\item{s_alpha_init}{initial value of proposal standard deviation for alpha's Metropolis step}

\item{a, b}{hyperparameters for alpha's prior}

\item{A, B}{square matrices of the hyperparameters for group-to-group probabilities' prior}

\item{n_swap}{number of swaps in order per iteration}

\item{N, thin, burnin, print_freq}{MCMC quantities}
}
\value{
a list of data frames and matrices representing the MCMC output
}
\description{
Run the regular Gibbs sampler of MMSBM
}
