% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vis_object}
\alias{vis_object}
\title{Obtain a standardised visNetwork object}
\usage{
vis_object(df.nodes, df.edges, layout = "layout_with_drl",
  select_by = "group", seed = 1234L, height = "720px")
}
\arguments{
\item{df.nodes}{data frame of nodes}

\item{df.edges}{data frame of edges}

\item{layout}{string of layout permitted by visNetwork::visIgraphLayout()}

\item{select_by}{string of variable to select by}

\item{seed}{random initial seed}

\item{height}{for height in visNetwork::visNetwork()}
}
\value{
a visNetwork object
}
\description{
Obtain a standardised visNetwork object
}
